organized cpp and hpp files
