--- Software/Arduino/motor_control.hpp
@@ -1,6 +1,0 @@
-#include <AccelStepper.h>
-
-class motorControl :: public AccelStepper {
-public: 
-
-};
